Refactor code to use SimilarBooksCubit in BookDetailsView
This commit refactors the code in the AppRouter and BookDetailsView files to use the SimilarBooksCubit in the BookDetailsView screen. A new route is defined in the AppRouter for the BookDetailsView screen, and the builder function creates a BlocProvider widget that provides the SimilarBooksCubit to the BookDetailsView widget as a child. In the BookDetailsView file, the SimilarBooksCubit is triggered to fetch similar books when the screen is initialized. Additionally, the SimilarBooksListView widget is updated to use the BlocBuilder and display a loading indicator or error message based on the state of the SimilarBooksCubit.