Add SimpleBlocObserver and HomeRepo
Added a new file, SimpleBlocObserver, which is a class that extends BlocObserver and provides methods to log changes, closures, and creations of Blocs and Cubits. This is useful for debugging state changes.

Also, edit, HomeRepo, Having a repository is a good practice to separate the data access logic from the presentation layer.